requirements.txt
Flask
firebase-admin
numpy
tensorflow==2.15.0
Pillow
APScheduler